Refactor NamespaceSelector to use GetMatches
This reduces some duplicated logic, and provides another example for
users of how to implement the ResourceList interface.
