+++ api/v1beta1/target.go
@@ -7,52 +7,11 @@ import (
 	"fmt"
 	"path/filepath"
 
-	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/client-go/dynamic"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
-
-// GetNamespaces fetches all namespaces in the cluster and returns a list of the
-// namespaces that match the NamespaceSelector. The client.Reader needs access
-// for viewing namespaces, like the access given by this kubebuilder tag:
-// `//+kubebuilder:rbac:groups=core,resources=namespaces,verbs=get;list;watch`
-func (sel NamespaceSelector) GetNamespaces(ctx context.Context, r client.Reader) ([]string, error) {
-	if len(sel.Include) == 0 && sel.LabelSelector == nil {
-		// A somewhat special case of no matches.
-		return []string{}, nil
-	}
-
-	listOpts := client.ListOptions{}
-
-	if sel.LabelSelector != nil {
-		labelSel, err := metav1.LabelSelectorAsSelector(sel.LabelSelector)
-		if err != nil {
-			return nil, err
-		}
-
-		listOpts.LabelSelector = labelSel
-	}
-
-	namespaceList := &corev1.NamespaceList{}
-	if err := r.List(ctx, namespaceList, &listOpts); err != nil {
-		return nil, err
-	}
-
-	namespaces := make([]string, len(namespaceList.Items))
-	for i, ns := range namespaceList.Items {
-		namespaces[i] = ns.GetName()
-	}
-
-	t := Target{
-		LabelSelector: sel.LabelSelector,
-		Include:       sel.Include,
-		Exclude:       sel.Exclude,
-	}
-
-	return t.matches(namespaces)
-}
 
 type Target struct {
 	*metav1.LabelSelector `json:",inline"`
@@ -174,32 +133,6 @@ func (t Target) matchesByName(items []client.Object) ([]client.Object, error) {
 	}
 
 	return matches, nil
-}
-
-// matches filters a slice of strings, and returns ones that match the Include
-// and Exclude lists in the Target. The only possible returned error is a
-// wrapped filepath.ErrBadPattern.
-func (t Target) matches(names []string) ([]string, error) {
-	// Using a map to ensure each entry in the result is unique.
-	set := make(map[string]struct{})
-
-	for _, name := range names {
-		matched, err := t.match(name)
-		if err != nil {
-			return nil, err
-		}
-
-		if matched {
-			set[name] = struct{}{}
-		}
-	}
-
-	matchingNames := make([]string, 0, len(set))
-	for ns := range set {
-		matchingNames = append(matchingNames, ns)
-	}
-
-	return matchingNames, nil
 }
 
 // match returns whether the given name matches the Include and Exclude lists in
